ProjectMachineLearning/app-engine-pipeline/requirements.txt
Flask==2.1.1
google-api-python-client==1.7.11
google-auth==1.25.0
google-auth-httplib2==0.0.3
google-cloud-language==1.3.0
google-cloud-speech==1.2.0
google-cloud-texttospeech==0.5.0
google-cloud-vision==0.39.0
googlemaps==3.1.3
gunicorn==19.9.0
pandas==1.1.5
requests==2.22.0
google-cloud-automl==2.7.2
markupsafe==2.0.1
jinja2==3.1.1
